replace linum-mode by display-line-number-mode
This  was done  in the  hope to  address the  problem with  screwed up
colors after loading another color  scheme which has unfortunately not
fixed now.  Obviously the  problem occurs only  if a  different scheme
then out of the  solarized group is used. When we  restrict to that we
should hopefully be able to live with that.

+ enable rainbow-params for scheme files

Signed-off-by: Otto Linnemann <linneman@gmx.de>